+++ internal/proposal/service.go
@@ -29,6 +29,7 @@ type DataProvider interface {
 	Update(proposal Proposal) error
 	GetByID(string) (*Proposal, error)
 	GetAvailableForVoting(time.Duration) ([]*Proposal, error)
+	GetByFilters(filters []Filter) (ProposalList, error)
 }
 
 type EventRegistered interface {
@@ -95,7 +96,7 @@ func (s *Service) processExisted(ctx context.Context, new, existed Proposal) err
 
 func (s *Service) checkSpecificUpdate(ctx context.Context, new, existed Proposal) {
 	if float64(new.ScoresTotal) >= new.Quorum {
-		go s.registerEventOnce(ctx, new, groupName, coreevents.SubjectProposalVotingReached)
+		go s.registerEventOnce(ctx, new, groupName, coreevents.SubjectProposalVotingQuorumReached)
 	}
 
 	if new.State != existed.State {
@@ -160,11 +161,29 @@ func (s *Service) processAvailableForVoting(ctx context.Context) error {
 			go s.registerEventOnce(ctx, *pr, groupName, coreevents.SubjectProposalVotingEnded)
 		}
 
-		// voting is coming
+		// voting will starts soon
 		if time.Now().Sub(startsAt) < startVotingWindow {
-			go s.registerEventOnce(ctx, *pr, groupName, coreevents.SubjectProposalVotingComing)
+			go s.registerEventOnce(ctx, *pr, groupName, coreevents.SubjectProposalVotingStartsSoon)
 		}
 	}
 
 	return nil
+}
+
+func (s *Service) GetByID(id string) (*Proposal, error) {
+	pro, err := s.repo.GetByID(id)
+	if err != nil {
+		return nil, fmt.Errorf("get by id: %w", err)
+	}
+
+	return pro, nil
+}
+
+func (s *Service) GetByFilters(filters []Filter) (ProposalList, error) {
+	list, err := s.repo.GetByFilters(filters)
+	if err != nil {
+		return ProposalList{}, fmt.Errorf("get by filters: %w", err)
+	}
+
+	return list, nil
 }