+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Dao gRPC server 
+- Proposal gRPC server 
+
 ## [0.0.4] - 2023-05-23
 
 ### Added